Balance info section updated
